2
Добавлена UI:
1. Рестарт уровня
2. Включение/выключение звука (пока звука нет и это просто заглушка)
3. Экран с сообщением о пройденном уровне с кнопкой начала следующего уровня

+++ Assets/Scripts/Field/Field.cs
@@ -12,7 +12,7 @@ public class Field : MonoBehaviour
     private Coroutine _shakeCorutine;
     private Coroutine _rotateCorutine;
     
-    public float BallsCount { get; private set; }
+    public int BallsCount { get; private set; }
 
     private void Awake()
     {
